requirements.txt
Flask[async]==2.2.2
PyYAML==6.0